Separated out class files for readability
